Add root CA creation
Finally add the creation of root certificates as well as the
necessary associated files to run a proper certificate authority
using the `openssl` command.

+++ openSsl.js
@@ -1,7 +1,10 @@
-const fs   = require("fs").promises;
-const path = require("path");
+const fs     = require("fs").promises;
+const path   = require("path");
 
 const child_process = require("mz/child_process");
+const crypto        = require("mz/crypto");
+
+const Config = require("./openSslConfig.js");
 
 
 function makeOpenSslFunction(subcommand)
@@ -12,6 +15,7 @@ function makeOpenSslFunction(subcommand)
 }
 
 const ca     = makeOpenSslFunction("ca");
+const genrsa = makeOpenSslFunction("genrsa");
 const req    = makeOpenSslFunction("req");
 const verify = makeOpenSslFunction("verify");
 const x509   = makeOpenSslFunction("x509");
@@ -41,8 +45,29 @@ module.exports = class OpenSsl
         }
     }
 
-    async withScratchDir(action)
+    /*!
+     * @breif Creates a temporary directory, which is passed to the @a action
+     *        function.
+     *
+     * @details
+     * This function creates a temporary scratch directory and then provides the
+     * path to the passed in @a action function. Once the @action function has
+     * completed the scratch directory is deleted.
+     *
+     * This is typically called as `withScratchDir((dir) => { doStuff; });`,
+     * however you can pass a boolean value as the first argument to controll
+     * whether or not the directory is delete afterwards:
+     * `withScratchDir(true, (dir) => { doStuff; });`.
+     */
+    async withScratchDir(action_or_keep, action=null)
     {
+        let keep = action_or_keep;
+        if (null == action)
+        {
+            keep = false;
+            action = action_or_keep
+        }
+
         let result = null;
 
         let scratch_dir = null
@@ -54,7 +79,7 @@ module.exports = class OpenSsl
         }
         finally
         {
-            if (scratch_dir)
+            if (scratch_dir && ! keep)
             {
                 await fs.rmdir(scratch_dir, {recursive: true});
             }
@@ -81,6 +106,78 @@ module.exports = class OpenSsl
             {
                 await file.close()
             }
+        }
+    }
+
+    async makeRootCert(name,
+                       key_length,
+                       digest_algorithm,
+                       lifetime,
+                       common_name,
+                       country,
+                       state,
+                       locality,
+                       organization,
+                       organizational_unit,
+                       email_address,
+                       intermediate_only,
+                       storage)
+    {
+        try
+        {
+            return await this.withScratchDir(async (scratch_dir) => {
+                const csr_conf_path    = path.join(scratch_dir, "csr.conf");
+                const key_path         = path.join(scratch_dir, "key");
+                const certificate_path = path.join(scratch_dir, "certificate");
+                const index_path       = path.join(scratch_dir, "index");
+                const serial_path      = path.join(scratch_dir, "serial");
+
+                await Promise.all([
+                    this.writeFile(csr_conf_path,
+                                   Config.getCreateCaConfig(digest_algorithm,
+                                                            common_name,
+                                                            country,
+                                                            state,
+                                                            locality,
+                                                            organization,
+                                                            organizational_unit,
+                                                            email_address)),
+
+                    // Create key pair
+                    genrsa("-out", key_path, key_length)
+                ]);
+
+                await Promise.all([
+                    // Create self-signed certificate
+                    req("-x509",
+                        "-new",
+                        "-config", csr_conf_path,
+                        "-extensions", "v3_ca",
+                        "-days", lifetime,
+                        "-key", key_path,
+                        "-out", certificate_path),
+
+                    // Create empty index file
+                    this.writeFile(index_path, ""),
+
+                    // Randomly initialize the serial number
+                    (crypto.randomBytes(2).then((serial) => {
+                        this.writeFile(serial_path, serial.toString('hex'));
+                    })),
+                ]);
+
+                return await storage.storeCert(name,
+                                               {certificate:       certificate_path,
+                                                key:               key_path,
+                                                index:             index_path,
+                                                serial:            serial_path},
+                                               {intermediate_only: intermediate_only});
+            });
+        }
+        catch (error)
+        {
+            console.log(error);
+            return false;
         }
     }
